+++ CHANGELOG.md
@@ -5,8 +5,11 @@ All notable changes to **oipd** will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 and the project follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.0]
+## [1.0.0] - 2025-07-10
 ### Added
+- overhauled user API interface. Now, users interface using the RND() class
+    - integrated plotting functionality
+    - integrated convenient function to return the probability future price will be at or above some value
 
 
 ## [0.0.6]